OGM-1235 move entities as nested classes in test
