+++ core/src/test/java/org/hibernate/ogm/backendtck/id/TableIdGeneratorInheritanceTest.java
@@ -6,35 +6,43 @@
  */
 package org.hibernate.ogm.backendtck.id;
 
+import java.util.UUID;
+import javax.persistence.Entity;
 import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Inheritance;
 
+import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.ogm.utils.jpa.OgmJpaTestCase;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import static javax.persistence.InheritanceType.TABLE_PER_CLASS;
 import static org.fest.assertions.Assertions.assertThat;
 
+/**
+ * Test that @Inheritance(strategy = TABLE_PER_CLASS) with generated id work correct
+ *
+ */
 public class TableIdGeneratorInheritanceTest extends OgmJpaTestCase {
 
+	private EntityManager em;
 
 	@Before
 	public void setUp() {
-
+		em = getFactory().createEntityManager();
 	}
 
 	@After
 	public void tearDown() {
-
+		em.close();
 	}
 
 	@Test
 	public void testTableIdGenerator() {
-		EntityManagerFactory emf = getFactory();
-		EntityManager em = emf.createEntityManager();
-
 		em.getTransaction().begin();
 		Truck loadedTruck = new Truck();
 		em.persist( loadedTruck );
@@ -49,19 +57,40 @@ public class TableIdGeneratorInheritanceTest extends OgmJpaTestCase {
 		assertThat( loadedTruck.getId() ).isNotNull();
 		assertThat( loadedTruck ).isNotNull();
 		em.getTransaction().commit();
-		em.close();
 	}
-
-//	@Override
-//	protected void configure(GetterPersistenceUnitInfo info) {
-//		TestHelper.enableCountersForInfinispan( info.getProperties() );
-//	}
 
 	@Override
 	public Class<?>[] getAnnotatedClasses() {
 		return new Class<?>[] {
 				BaseCar.class,
-				Truck.class
+				Truck.class,
+				PassengerCar.class
 		};
 	}
+
+	@Entity
+	private static class Truck extends BaseCar {
+	}
+
+	@Entity
+	private static class PassengerCar extends BaseCar {
+	}
+
+	@Entity
+	@Inheritance(strategy = TABLE_PER_CLASS)
+	private abstract static class BaseCar {
+		protected UUID id;
+
+		@Id
+		@GeneratedValue(generator = "uuid")
+		@GenericGenerator(name = "uuid", strategy = "uuid2")
+		public UUID getId() {
+			return id;
+		}
+
+		public void setId(UUID id) {
+			this.id = id;
+		}
+	}
+
 }
