--- core/src/test/java/org/hibernate/ogm/backendtck/id/Truck.java
@@ -1,26 +1,0 @@
-/*
- * Hibernate OGM, Domain model persistence for NoSQL datastores
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.ogm.backendtck.id;
-
-import javax.persistence.Entity;
-
-@Entity
-public class Truck extends BaseCar {
-//	protected UUID id;
-//
-//	@Id
-//	// @GeneratedValue  //<--tried this only this one as well
-//	@GeneratedValue(generator = "uuid")
-//	@GenericGenerator(name = "uuid", strategy = "uuid2")
-//	public UUID getId() {
-//		return id;
-//	}
-//
-//	public void setId(UUID id) {
-//		this.id = id;
-//	}
-}
